Removed native hota code
