--- src/av2/evaluation/scenario_mining/metrics.py
@@ -1,8 +1,0 @@
-from av2.evaluation.scenario_mining.hota import HOTA
-from trackeval.metrics.clear import CLEAR
-from trackeval.metrics.identity import Identity
-from trackeval.metrics.count import Count
-from trackeval.metrics.j_and_f import JAndF
-from trackeval.metrics.track_map import TrackMAP
-from trackeval.metrics.vace import VACE
-from trackeval.metrics.ideucl import IDEucl
